refactor alias model integer->float, loancalculator stuff

+++ src/main/java/fi/metropolia/group8/view/PrimaryController.java
@@ -97,7 +97,7 @@ public class PrimaryController {
         if(aliasDataModel.getCurrentAlias() != null) {
             try {
                 primaryCurrentAlias.setText(aliasDataModel.getCurrentAlias().getName());
-                primaryCurrentEquity.setText(Integer.toString(aliasDataModel.getCurrentAlias().getEquity()));
+                primaryCurrentEquity.setText(Float.toString(aliasDataModel.getCurrentAlias().getEquity()));
             }
             catch (Exception e) {
                 e.printStackTrace();
